Instruction change following UI Update.
Datapoint y coordinate changed to double.

+++ Corona Forecast/src/com/Corona/Forecast/Controller/DataPoint.java
@@ -2,7 +2,7 @@ package com.Corona.Forecast.Controller;
 
 public class DataPoint {
 	public int x; 
-	public int y; 
+	public double y; 
 	 
 	 public DataPoint()
 	 {
